requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0

# Document loading
pdfplumber>=0.11.0
python-docx>=1.1.2

# Embeddings
sentence-transformers>=3.0.0
torch>=2.0.0

# Vector store
faiss-cpu>=1.8.0

# Hybrid search / reranker
numpy>=1.26.0

# Utilities
python-multipart>=0.0.9
aiofiles>=23.2.1
PyYAML>=6.0.1
